Add version-specific casks for each major/minor series of Docker CE releases.
Each of these version-specific casks corresponds to the most recent stable
release in that major/minor series. Going forward, the plan is to keep each
of these casks aligned with the latest release in that major/minor series.
For example, there were three releases in the 17.03 series: 17.03.0-ce-mac1,
17.03.1-ce-mac5, and 17.03.1-ce-mac12. If "docker@17.03-ce.rb" had been
introduced at the start of that series, it would have been committed at
least three times, once for each release within the 17.03 series.